+++ jsmetrics/windspeed_utils.py
@@ -1,8 +1,8 @@
 # -*- coding: utf-8 -*-
 
 """
-    Operations needed for the jet-stream metrics and algorithms that specifically operate on windspeed data.
-    Includes the base classes and function for dealing with windspeed vectors and slices of windspeed
+    Operations needed for the jet-stream metrics and jet-stream algorithms that specifically operate on windspeed data.
+    Includes the base classes and function for dealing with windspeed vectors and lat/lon or lat/plev slices of windspeed (so called: WindSpeedSlice class)
 """
 
 # imports