+++ HISTORY.rst
@@ -75,7 +75,7 @@ BIG CHANGES:
 
 0.0.3-beta (2022-01-14)
 ------------------
-* Use real part from fourier filter to woolings and its tests
+* Use real part from fourier filter to Woollings and its tests
 
 0.0.3-alpha (2022-01-14)
 ------------------